+++ src/safejax/__init__.py
@@ -3,5 +3,5 @@
 __author__ = "Alvaro Bartolome <alvarobartt@yahoo.com>"
 __version__ = "0.3.0"
 
-from safejax.load import deserialize  # noqa: F401
-from safejax.save import serialize  # noqa: F401
+from safejax.core.load import deserialize  # noqa: F401
+from safejax.core.save import serialize  # noqa: F401
