+++ src/safejax/core/load.py
@@ -1,4 +1,3 @@
-import os
 from pathlib import Path
 from typing import Union
 
@@ -6,7 +5,7 @@ from flax.core.frozen_dict import freeze
 from objax.variable import VarCollection
 from safetensors.flax import load, load_file
 
-from safejax.typing import DictionaryLike, PathLike
+from safejax.typing import ParamsDictLike, PathLike
 from safejax.utils import unflatten_dict
 
 
@@ -15,7 +14,7 @@ def deserialize(
     freeze_dict: bool = False,
     requires_unflattening: bool = True,
     to_var_collection: bool = False,
-) -> DictionaryLike:
+) -> ParamsDictLike:
     """
     Deserialize JAX, Flax, Haiku, or Objax model params from either a `bytes` object or a file path,
     stored using `safetensors.flax.save_file` or directly saved using `safejax.save.serialize` with
@@ -43,12 +42,18 @@ def deserialize(
     """
     if isinstance(path_or_buf, bytes):
         decoded_params = load(data=path_or_buf)
-    if (
-        isinstance(path_or_buf, str)
-        or isinstance(path_or_buf, Path)
-        or isinstance(path_or_buf, os.PathLike)
-    ):
-        decoded_params = load_file(filename=path_or_buf)
+    elif isinstance(path_or_buf, (str, Path)):
+        filename = path_or_buf if isinstance(path_or_buf, Path) else Path(path_or_buf)
+        if not filename.exists or not filename.is_file:
+            raise ValueError(
+                f"`path_or_buf` must be a valid file path, not {path_or_buf}."
+            )
+        decoded_params = load_file(filename=filename.as_posix())
+    else:
+        raise ValueError(
+            "`path_or_buf` must be a `bytes` object or a file path (`str` or"
+            f" `pathlib.Path` object), not {type(path_or_buf)}."
+        )
     if requires_unflattening:
         decoded_params = unflatten_dict(params=decoded_params)
     if freeze_dict:
